+++ R/qm_rasterToPolygons.r
@@ -4,9 +4,9 @@
 #' It's faster by turning off the checking done in the simple features package, but it's also faster
 #' than raster because it uses a dense mesh to generate the coordinates. 
 #'
-#' Note that `na.rm` only applies to the first layer, get in touch if this is a problem. 
 #' @param x raster, brick or stack
-#' @param na.rm defaults to `FALSE`, if `TRUE` will polygonize only the populated cells of the first layer
+#' @param na.rm defaults to `TRUE` and will polygonize only the cells that are non-NA across all layers, 
+#' set to `FALSE` to not remove any cells
 #' @param ... arguments passed to methods, currently unused
 #'
 #' @return simple features POLYGON layer, or SpatialPolygonsDataFrame
@@ -37,7 +37,19 @@
 #' @name polygonize
 #' @aliases qm_rasterToPolygons qm_rasterToPolygons_sp
 #' @export
-polygonize.RasterLayer <- function(x, na.rm = FALSE, ...) {
+polygonize.RasterLayer <- function(x, na.rm = TRUE, ...) {
+  ## get all the layers off the raster
+  sf1 <- stats::setNames(as.data.frame(raster::values(x)), names(x))
+  
+  na_rm <- na.rm
+  if (na.rm && raster::nlayers(x) > 1) {
+    na_all <- Reduce(`&`, lapply(sf1, function(x) is.na(x)))
+    na.rm <- FALSE
+  } else {
+    na_all <- is.na(sf1[[1]])
+    
+  }
+  
   ## create dense mesh of cell corner coordinates
   qm <- quadmesh::quadmesh(x, z = NULL, na.rm = na.rm)
   ## a dummy structure to copy
@@ -46,20 +58,18 @@ polygonize.RasterLayer <- function(x, na.rm = FALSE, ...) {
   
   ## TODO: speed up, this is the slow part
   spl <- split(t(qm$vb[1:2, qm$ib]), rep(seq_len(ncol(qm$ib)), each = 4))
+  ## remove the *common-missing* quads here
+  if (na_rm) {
+    sf1 <- sf1[!na_all, , drop = FALSE]
+    spl <- spl[!na_all]
+    
+  }
   l <- lapply(spl, function(a) {
     template[[1L]] <- 
     cbind(a[c(1, 2, 3, 4, 1)], a[c(5, 6, 7, 8, 5)])
     template
     })
   
-  ## get all the layers off the raster
-  sf1 <- stats::setNames(as.data.frame(raster::values(x)), names(x))
-  
-  if (na.rm) {
-    sf1 <- sf1[!is.na(sf1[[1]]), , drop = FALSE]
-  }
-  ## add the geometry column
-  #sf1[["geometry"]] <- sf::st_sfc(l)
   ex <- extent(x)
   sf1[["geometry"]] <- structure(l, n_empty = 0L, 
                                  crs = structure(list(epsg = NA_integer_, proj4string = raster::projection(x)), class = "crs"),
@@ -89,7 +99,7 @@ polygonize.RasterBrick <- qm_rasterToPolygons
 #' @name polygonize
 #' @export
 #' @importFrom raster getValues
-qm_rasterToPolygons_sp <- function(x, na.rm = FALSE, ...) {
+qm_rasterToPolygons_sp <- function(x, na.rm = TRUE, ...) {
   x0 <- polygonize(x, na.rm = na.rm)
   g <- unclass(x0[[attr(x0, "sf_column")]])
   x0[[attr(x0, "sf_column")]] <- NULL